Move compare_obj_to_small test next to similar tests
